Updated code

+++ src/main/java/example/algo/twopointers/example/ThreeSum.java
@@ -44,6 +44,6 @@ public class ThreeSum {
                 }
             }
         }
-        return minimSum;
+        return (int)minimSum;
     }
 }
